Hexen: add support for ZDBSP and DeePBSP nodes

+++ src/hexen/p_local.h
@@ -307,8 +307,8 @@ extern void P_SegLengths (void);
 extern void P_RestoreSectorBrightness (void);
 
 extern byte *rejectmatrix;      // for fast sight rejection
-extern short *blockmaplump;     // offsets in blockmap are from here
-extern short *blockmap;
+extern int32_t *blockmaplump;     // offsets in blockmap are from here
+extern int32_t *blockmap;
 extern int bmapwidth, bmapheight;       // in mapblocks
 extern fixed_t bmaporgx, bmaporgy;      // origin of block map
 extern mobj_t **blocklinks;     // for thing chains
