+++ src/hexen/p_maputl.c
@@ -428,7 +428,7 @@ If the function returns false, exit with false without checking anything else.
 boolean P_BlockLinesIterator(int x, int y, boolean(*func) (line_t *))
 {
     int offset;
-    short *list;
+    int32_t *list;
     line_t *ld;
 
     int i;
