+++ src/hexen/p_sight.c
@@ -90,7 +90,7 @@ static boolean PTR_SightTraverse(intercept_t * in)
 static boolean P_SightBlockLinesIterator(int x, int y)
 {
     int offset;
-    short *list;
+    int32_t *list;
     line_t *ld;
     int s1, s2;
     divline_t dl;
